# DDD - Falta Print das Telas/Documentação API/Raízes Solidárias.postman_collection.json-OpenApi3Json.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Raízes Solidárias",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "variables": {}
    },
    {
      "url": "https://localhost:8080",
      "variables": {}
    }
  ],
  "paths": {
...
    "/destino": {
      "get": {
        "tags": [
          "Destino"
        ],
        "summary": "Listar Destinos",
        "operationId": "ListarDestinos",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      },
      "post": {
        "tags": [
          "Destino"
        ],
        "summary": "Cadastrar Destino",
        "operationId": "CadastrarDestino",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CadastrarDestinoRequest"
                  },
                  {
                    "example": {
                      "endereco_destino": "Endereço do destino",
                      "responsavel_destino": "Responsável pelo destino",
                      "cel_destino": "(11) 98309-0659",
                      "qtd_dependentes_destino": 5
                    }
                  }
                ]
              },
              "example": {
                "endereco_destino": "Endereço do destino",
                "responsavel_destino": "Responsável pelo destino",
                "cel_destino": "(11) 98309-0659",
                "qtd_dependentes_destino": 5
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
    "/destino/12": {
      "put": {
        "tags": [
          "Destino"
        ],
        "summary": "Atualizar Destino",
        "operationId": "AtualizarDestino",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AtualizarDestinoRequest"
                  },
                  {
                    "example": {
                      "id_destino": 12,
                      "endereco_destino": "Endereço do destino",
                      "responsavel_destino": "Responsável pelo sDASDASDASDASD",
                      "cel_destino": "(11) 98309-0659",
                      "qtd_dependentes_destino": 1000000
                    }
                  }
                ]
              },
              "example": {
                "id_destino": 12,
                "endereco_destino": "Endereço do destino",
                "responsavel_destino": "Responsável pelo sDASDASDASDASD",
                "cel_destino": "(11) 98309-0659",
                "qtd_dependentes_destino": 1000000
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      },
      "delete": {
        "tags": [
          "Destino"
        ],
        "summary": "Deletar Destino",
        "operationId": "DeletarDestino",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "https://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
    "/plantio": {
      "get": {
        "tags": [
          "Plantio"
        ],
        "summary": "Listar Plantios",
        "operationId": "ListarPlantios",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      },
      "post": {
        "tags": [
          "Plantio"
        ],
        "summary": "Cadastrar Plantio",
        "operationId": "CadastrarPlantio",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CadastrarPlantioRequest"
                  },
                  {
                    "example": {
                      "data_plantio": "2023-06-01",
                      "espaco_plantio": 10,
                      "alimento": {
                        "id_alimento": 1
                      }
                    }
                  }
                ]
              },
              "example": {
                "data_plantio": "2023-06-01",
                "espaco_plantio": 10,
                "alimento": {
                  "id_alimento": 1
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
    "/plantio/12": {
      "put": {
        "tags": [
          "Plantio"
        ],
        "summary": "Atualizar Plantio",
        "operationId": "AtualizarPlantio",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AtualizarPlantioRequest"
                  },
                  {
                    "example": {
                      "id_plantio": 12,
                      "data_plantio": "2023-06-30",
                      "espaco_plantio": 20,
                      "alimento": {
                        "id_alimento": 2
                      }
                    }
                  }
                ]
              },
              "example": {
                "id_plantio": 12,
                "data_plantio": "2023-06-30",
                "espaco_plantio": 20,
                "alimento": {
                  "id_alimento": 2
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
    "/plantio/plantios": {
      "post": {
        "tags": [
          "Plantio"
        ],
        "summary": "Cadastrar Plantios",
        "operationId": "CadastrarPlantios",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CadastrarPlantiosRequest"
                },
                "description": "",
                "example": [
                  {
                    "data_plantio": "2023-06-01",
                    "espaco_plantio": 10,
                    "alimento": {
                      "id_alimento": 1
                    }
                  },
                  {
                    "data_plantio": "2023-06-02",
                    "espaco_plantio": 8,
                    "alimento": {
                      "id_alimento": 2
                    }
                  }
                ]
              },
              "example": [
                {
                  "data_plantio": "2023-06-01",
                  "espaco_plantio": 10,
                  "alimento": {
                    "id_alimento": 1
                  }
                },
                {
                  "data_plantio": "2023-06-02",
                  "espaco_plantio": 8,
                  "alimento": {
                    "id_alimento": 2
                  }
                }
              ]
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AtualizarDestinoRequest": {
        "title": "AtualizarDestinoRequest",
        "required": [
          "id_destino",
          "endereco_destino",
          "responsavel_destino",
          "cel_destino",
          "qtd_dependentes_destino"
        ],
        "type": "object",
        "properties": {
          "id_destino": {
            "type": "integer",
            "format": "int32"
          },
          "endereco_destino": {
            "type": "string"
          },
          "responsavel_destino": {
            "type": "string"
          },
          "cel_destino": {
            "type": "string"
          },
          "qtd_dependentes_destino": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "id_destino": 12,
          "endereco_destino": "Endereço do destino",
          "responsavel_destino": "Responsável pelo sDASDASDASDASD",
          "cel_destino": "(11) 98309-0659",
          "qtd_dependentes_destino": 1000000
        }
      },
      "CadastrarDestinoRequest": {
        "title": "CadastrarDestinoRequest",
        "required": [
          "endereco_destino",
          "responsavel_destino",
          "cel_destino",
          "qtd_dependentes_destino"
        ],
        "type": "object",
        "properties": {
          "endereco_destino": {
            "type": "string"
          },
          "responsavel_destino": {
            "type": "string"
          },
          "cel_destino": {
            "type": "string"
          },
          "qtd_dependentes_destino": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "endereco_destino": "Endereço do destino",
          "responsavel_destino": "Responsável pelo destino",
          "cel_destino": "(11) 98309-0659",
          "qtd_dependentes_destino": 5
        }
      },
...
      "AtualizarPlantioRequest": {
        "title": "AtualizarPlantioRequest",
        "required": [
          "id_plantio",
          "data_plantio",
          "espaco_plantio",
          "alimento"
        ],
        "type": "object",
        "properties": {
          "id_plantio": {
            "type": "integer",
            "format": "int32"
          },
          "data_plantio": {
            "type": "string"
          },
          "espaco_plantio": {
            "type": "integer",
            "format": "int32"
          },
          "alimento": {
            "$ref": "#/components/schemas/Alimento"
          }
        },
        "example": {
          "id_plantio": 12,
          "data_plantio": "2023-06-30",
          "espaco_plantio": 20,
          "alimento": {
            "id_alimento": 2
          }
        }
      },
      "Alimento": {
        "title": "Alimento",
        "required": [
          "id_alimento"
        ],
        "type": "object",
        "properties": {
          "id_alimento": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "id_alimento": 2
        }
      },
      "CadastrarPlantioRequest": {
        "title": "CadastrarPlantioRequest",
        "required": [
          "data_plantio",
          "espaco_plantio",
          "alimento"
        ],
        "type": "object",
        "properties": {
          "data_plantio": {
            "type": "string"
          },
          "espaco_plantio": {
            "type": "integer",
            "format": "int32"
          },
          "alimento": {
            "$ref": "#/components/schemas/Alimento"
          }
        },
        "example": {
          "data_plantio": "2023-06-01",
          "espaco_plantio": 10,
          "alimento": {
            "id_alimento": 1
          }
        }
      },
      "CadastrarPlantiosRequest": {
        "title": "CadastrarPlantiosRequest",
        "required": [
          "data_plantio",
          "espaco_plantio",
          "alimento"
        ],
        "type": "object",
        "properties": {
          "data_plantio": {
            "type": "string"
          },
          "espaco_plantio": {
            "type": "integer",
            "format": "int32"
          },
          "alimento": {
            "$ref": "#/components/schemas/Alimento"
          }
        },
        "example": {
          "data_plantio": "2023-06-01",
          "espaco_plantio": 10,
          "alimento": {
            "id_alimento": 1
          }
        }
      },
...
    }
  },
...
}